requirements.txt
aniso8601==9.0.1
blinker==1.7.0
certifi==2023.7.22
cffi==1.16.0
click==8.1.7
cloudinary==1.36.0
colorama==0.4.6
cryptography==41.0.5
Flask==3.0.0
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
PyMySQL==1.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
WTForms==3.1.1
